+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This file is used to list changes made in each version of the aws-parallelcluste
 - Add possibility to override EC2 RunInstances parameters for instances launched in a Slurm cluster.
 
 **CHANGES**
-- Upgrade Slurm to version 21.08.3.
+- Update Slurm plugin to support version 21.08.
 
 3.0.2
 ------